separate level buttons component
