+++ package.json
@@ -8,14 +8,18 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
+    "@mediapipe/camera_utils": "^0.3.1640029074",
+    "@mediapipe/drawing_utils": "^0.3.1620248257",
+    "@mediapipe/hands": "^0.4.1646424915",
     "@react-three/drei": "^9.5.0",
     "@react-three/fiber": "^8.0.12",
     "autoprefixer": "^10.4.4",
     "gsap": "^3.10.3",
     "jotai": "^1.6.6",
     "postcss": "^8.4.12",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-webcam": "^7.0.1",
     "tailwindcss": "^3.0.24",
     "three": "^0.139.2"
   },
